Refactor
Change foldername to match lobby/master
