+++ source/model/Room.js
@@ -1,4 +1,4 @@
 /**
- * @module models/Room
+ * @module model/Room
  * @type {*|exports|module.exports}
  */
